api/requirements.txt
torch>=2.3.0
torchvision>=0.18.0
transformers>=4.44.0
Pillow>=10.0.0
numpy>=1.26.0
fastapi>=0.111.0
uvicorn>=0.30.0
pydantic>=2.8.0
python-multipart